add input and output handler class
Added a CodecConverter for both encoding and decoding the txt file, as well as created a dataHandler to communicate between controller and the codecConverter. Encapsulated both the model and datahandler with a singleton pattern to just make sure there's only a single instance of them, might not be needed fully at the moment however still a good precaution for future updates.